+++ plugins/modules/oci_ai_document_processor_job_actions.py
@@ -101,6 +101,18 @@ processor_job:
                     returned: on success
                     type: complex
                     contains:
+                        model_id:
+                            description:
+                                - The custom model ID.
+                            returned: on success
+                            type: str
+                            sample: "ocid1.model.oc1..xxxxxxEXAMPLExxxxxx"
+                        tenancy_id:
+                            description:
+                                - The custom model tenancy ID when modelId represents aliasName.
+                            returned: on success
+                            type: str
+                            sample: "ocid1.tenancy.oc1..xxxxxxEXAMPLExxxxxx"
                         max_results:
                             description:
                                 - The maximum number of results to return.
@@ -251,6 +263,8 @@ processor_job:
             "processor_type": "GENERAL",
             "document_type": "INVOICE",
             "features": [{
+                "model_id": "ocid1.model.oc1..xxxxxxEXAMPLExxxxxx",
+                "tenancy_id": "ocid1.tenancy.oc1..xxxxxxEXAMPLExxxxxx",
                 "max_results": 56,
                 "feature_type": "LANGUAGE_CLASSIFICATION",
                 "generate_searchable_pdf": true
@@ -299,7 +313,7 @@ except ImportError:
     HAS_OCI_PY_SDK = False
 
 
-class ProcessorJobActionsHelperGen(OCIActionsHelperBase):
+class AiDocumentProcessorJobActionsHelperGen(OCIActionsHelperBase):
     """
     Supported actions:
         cancel
@@ -341,10 +355,14 @@ class ProcessorJobActionsHelperGen(OCIActionsHelperBase):
         )
 
 
-ProcessorJobActionsHelperCustom = get_custom_class("ProcessorJobActionsHelperCustom")
+AiDocumentProcessorJobActionsHelperCustom = get_custom_class(
+    "AiDocumentProcessorJobActionsHelperCustom"
+)
 
 
-class ResourceHelper(ProcessorJobActionsHelperCustom, ProcessorJobActionsHelperGen):
+class ResourceHelper(
+    AiDocumentProcessorJobActionsHelperCustom, AiDocumentProcessorJobActionsHelperGen
+):
     pass